Version 1.04g
Fixed (hopefully) issue with occasional very high Z probe readings when
using a simple modulated sensor

+++ Platform.cpp
@@ -374,7 +374,7 @@ int Platform::ZProbe() const
 			// We assume that zProbeOnFilter and zProbeOffFilter average the same number of readings.
 			// Because of noise, it is possible to get a negative reading, so allow for this.
 			return (int) (((int32_t) zProbeOnFilter.GetSum() - (int32_t) zProbeOffFilter.GetSum())
-					/ (4 * numZProbeReadingsAveraged));
+					/ (int)(4 * numZProbeReadingsAveraged));
 
 		default:
 			break;
